Add files via upload
This project is a comprehensive solution for managing menu-based applications using a variety of technologies. It integrates Python for core functionality, Docker for containerization, AWS for cloud deployment, and Linux for seamless environment management. The aim of the project is to streamline the deployment and scaling of menu-driven applications with an emphasis on portability and scalability.

#Python #Docker #AWS #Linux #CloudComputing #Automation #MenuBase #OpenSource #DevOps #Project#Menubase #Portfolio